Revert "Revert "add /Users/jonathan.homan/bin to PATH; add kubectx, kubens""
This reverts commit b21e8d1178cae0d60a6dd57dbfd54fc348bcd53a.

Okay, so let's bring this back. Installing a specific version of
kubernetes-cli is not straightforward in homebrew. For now, install
kubectx and kubens manually and rely on the kubectl from Docker. In the
future, I think I'll add another step to download the desired version of
kubectl from GH, etc.